backend/requirements.txt
asgiref==3.7.2
black==23.12.1
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.7
defusedxml==0.8.0rc2
diff-match-patch==20230430
Django==3.2.3
django-advanced-filters @ https://github.com/modlinltd/django-advanced-filters/archive/3fc6b964c48c38fff6169f55f2c26325aa926dfb.zip#sha256=6723130e063988b774d0d4de044e5881e1792bb9aa6be6474f401ef097c3c597
django-braces==1.14.0
django-colorfield==0.11.0
django-cors-headers==3.13.0
django-filter==23.3
django-import-export==3.3.6
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
et-xmlfile==1.1.0
flake8==6.1.0
flake8-isort==6.1.1
gunicorn==20.1.0
idna==3.6
isort==5.13.2
itypes==1.2.0
Jinja2==3.1.2
loguru==0.7.2
MarkupPy==1.14
MarkupSafe==2.1.3
mccabe==0.7.0
mypy-extensions==1.0.0
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.2
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.1.0
psycopg2-binary==2.9.3
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.1.0
PyJWT==2.8.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
reportlab==4.1.0
requests==2.31.0
requests-oauthlib==1.3.1
simplejson==3.19.2
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.5.1
sqlparse==0.4.4
tablib==3.5.0
tomli==2.0.1
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.1.0
xlrd==2.0.1
xlwt==1.3.0
